requirements.txt
streamlit
rdkit
plotly
pandas
pillow